refactor: move instant-trade types to types file
