+++ packages/sdk/src/index.ts
@@ -30,7 +30,7 @@ export * as unisat from "./browser-wallets/unisat"
 export * as xverse from "./browser-wallets/xverse"
 export * from "./config"
 export * from "./inscription"
-export { InstantTradeBuilder, InstantTradeBuyerTxBuilder, InstantTradeSellerTxBuilder } from "./instant-trade"
+export * from "./instant-trade"
 export * from "./keys"
 export { BaseDatasource, JsonRpcDatasource } from "./modules"
 export * from "./signatures"
