+++ packages/sdk/src/instant-trade/index.ts
@@ -1,3 +1,4 @@
 export { default as InstantTradeBuilder } from "./InstantTradeBuilder"
 export { default as InstantTradeBuyerTxBuilder } from "./InstantTradeBuyerTxBuilder"
 export { default as InstantTradeSellerTxBuilder } from "./InstantTradeSellerTxBuilder"
+export * from "./types"
